requirements.txt
aiogram==2.25.1
gspread==5.12.0
oauth2client==4.1.3
APScheduler==3.10.1
google-api-python-client==2.126.0
python-dotenv
PySide6==6.7.2
requests==2.32.3
